rename EndToEnd -> EndToEndSingle
