+++ examples/tabarena/run_beta_tabpfn_end_to_end.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 from tabrepo.nips2025_utils.artifacts.download_utils import download_and_extract_zip
-from tabrepo.nips2025_utils.end_to_end import EndToEnd, EndToEndResults
+from tabrepo.nips2025_utils.end_to_end import EndToEndSingle, EndToEndResultsSingle
 
 
 if __name__ == '__main__':
@@ -31,7 +31,7 @@ if __name__ == '__main__':
     Once this is executed once, it does not need to be ran again.
     """
     if cache:
-        end_to_end = EndToEnd.from_path_raw(path_raw=path_raw)
+        end_to_end = EndToEndSingle.from_path_raw(path_raw=path_raw)
 
     """
     Load cached results and compare on TabArena
@@ -39,6 +39,6 @@ if __name__ == '__main__':
     2. Currently compares on all datasets, does not compare on subsets.
     3. Missing values are imputed to default RandomForest.
     """
-    end_to_end_results = EndToEndResults.from_cache(method=method)
+    end_to_end_results = EndToEndResultsSingle.from_cache(method=method)
     leaderboard = end_to_end_results.compare_on_tabarena(output_dir=fig_output_dir)
     print(leaderboard)
