+++ tabrepo/nips2025_utils/end_to_end.py
@@ -1,19 +1,13 @@
 from __future__ import annotations
 
 from pathlib import Path
-from typing import TYPE_CHECKING
-from typing_extensions import Self
 
 import pandas as pd
 
-from tabrepo.benchmark.result import BaselineResult, ConfigResult
-from tabrepo.nips2025_utils.artifacts.method_metadata import MethodMetadata
+from tabrepo.benchmark.result import BaselineResult
 from tabrepo.nips2025_utils.compare import compare_on_tabarena
-from tabrepo.nips2025_utils.method_processor import get_info_from_result, generate_task_metadata, load_raw
-from tabrepo.nips2025_utils.tabarena_context import TabArenaContext
-
-if TYPE_CHECKING:
-    from tabrepo.repository import EvaluationRepository
+from tabrepo.nips2025_utils.end_to_end_single import EndToEndSingle, EndToEndResultsSingle
+from tabrepo.nips2025_utils.method_processor import get_info_from_result
 
 
 class EndToEndMulti:
@@ -24,7 +18,7 @@ class EndToEndMulti:
         cache: bool = True,
     ):
         # raw
-        self.results_lst: list[BaselineResult] = EndToEnd.clean_raw(results_lst=results_lst)
+        self.results_lst: list[BaselineResult] = EndToEndSingle.clean_raw(results_lst=results_lst)
 
         result_types_dict = {}
         for r in self.results_lst:
@@ -40,7 +34,7 @@ class EndToEndMulti:
         for cur_type in unique_types:
             # TODO: Avoid fetching task_metadata repeatedly from OpenML in each iteration
             cur_results_lst = result_types_dict[cur_type]
-            cur_end_to_end = EndToEnd(results_lst=cur_results_lst, task_metadata=task_metadata, cache=cache)
+            cur_end_to_end = EndToEndSingle(results_lst=cur_results_lst, task_metadata=task_metadata, cache=cache)
             self.end_to_end_lst.append(cur_end_to_end)
 
     def to_results(self) -> EndToEndResultsMulti:
@@ -49,85 +43,10 @@ class EndToEndMulti:
         )
 
 
-class EndToEnd:
-    def __init__(
-        self,
-        results_lst: list[BaselineResult | dict],
-        task_metadata: pd.DataFrame | None = None,
-        cache: bool = True,
-    ):
-        # raw
-        self.results_lst: list[BaselineResult] = self.clean_raw(results_lst=results_lst)
-        self.method_metadata: MethodMetadata = MethodMetadata.from_raw(
-            results_lst=self.results_lst
-        )
-
-        if cache:
-            print(
-                f'Artifacts for method {self.method_metadata.method} will be saved in: "{self.method_metadata.path}"'
-            )
-
-        if cache:
-            self.method_metadata.to_yaml()
-            self.method_metadata.cache_raw(results_lst=self.results_lst)
-
-        if task_metadata is None:
-            tids = list({r.task_metadata["tid"] for r in self.results_lst})
-            task_metadata = generate_task_metadata(tids=tids)
-        self.task_metadata = task_metadata
-
-        # processed
-        self.repo: EvaluationRepository = self.method_metadata.generate_repo(
-            results_lst=self.results_lst,
-            task_metadata=self.task_metadata,
-            cache=cache,
-        )
-
-        # results
-        tabarena_context = TabArenaContext()
-        self.hpo_results, self.model_results = tabarena_context.simulate_repo(
-            method=self.method_metadata,
-            repo=self.repo,
-            use_rf_config_fallback=False,
-            cache=cache,
-        )
-
-    @classmethod
-    def clean_raw(cls, results_lst: list[BaselineResult | dict]) -> list[BaselineResult]:
-        return [r if not isinstance(r, dict) else BaselineResult.from_dict(result=r) for r in results_lst]
-
-    @classmethod
-    def from_path_raw(cls, path_raw: str | Path, name: str = None, name_suffix: str = None) -> Self:
-        results_lst: list[BaselineResult] = load_raw(path_raw=path_raw)
-        if name is not None or name_suffix is not None:
-            for r in results_lst:
-                r.update_name(name=name, name_suffix=name_suffix)
-                if isinstance(r, ConfigResult):
-                    r.update_model_type(name=name, name_suffix=name_suffix)
-        return cls(results_lst=results_lst)
-
-    @classmethod
-    def from_cache(cls, method: str, artifact_name: str | None = None) -> Self:
-        if artifact_name is None:
-            artifact_name = method
-        method_metadata = MethodMetadata.from_yaml(
-            method=method, artifact_name=artifact_name
-        )
-        results_lst = method_metadata.load_raw()
-        return cls(results_lst=results_lst)
-
-    def to_results(self) -> EndToEndResults:
-        return EndToEndResults(
-            method_metadata=self.method_metadata,
-            hpo_results=self.hpo_results,
-            model_results=self.model_results,
-        )
-
-
 class EndToEndResultsMulti:
     def __init__(
         self,
-        end_to_end_results_lst: list[EndToEndResults],
+        end_to_end_results_lst: list[EndToEndResultsSingle],
     ):
         self.end_to_end_results_lst = end_to_end_results_lst
 
@@ -180,110 +99,3 @@ class EndToEndResultsMulti:
             ))
         df_results = pd.concat(df_results_lst, ignore_index=True)
         return df_results
-
-
-class EndToEndResults:
-    def __init__(
-        self,
-        method_metadata: MethodMetadata,
-        hpo_results: pd.DataFrame = None,
-        model_results: pd.DataFrame = None,
-    ):
-        self.method_metadata = method_metadata
-        if hpo_results is None and self.method_metadata.method_type == "config":
-            hpo_results = self.method_metadata.load_hpo_results()
-        if model_results is None:
-            model_results = self.method_metadata.load_model_results()
-        self.hpo_results = hpo_results
-        self.model_results = model_results
-
-    @classmethod
-    def from_cache(cls, method: str, artifact_name: str | None = None) -> Self:
-        if artifact_name is None:
-            artifact_name = method
-        method_metadata = MethodMetadata.from_yaml(
-            method=method, artifact_name=artifact_name
-        )
-        return cls(method_metadata=method_metadata)
-
-    def compare_on_tabarena(
-        self,
-        output_dir: str | Path,
-        *,
-        filter_dataset_fold: bool = False,
-        subset: str | None | list = None,
-        new_result_prefix: str | None = None,
-        use_model_results: bool = False,
-    ) -> pd.DataFrame:
-        """Compare results on TabArena leaderboard.
-
-        Args:
-            output_dir (str | Path): Directory to save the results.
-            subset (str | None | list): Subset of tasks to evaluate on.
-                Options are "classification", "regression", "lite"  for TabArena-Lite,
-                "tabicl", "tabpfn", "tabpfn/tabicl", or None for all tasks.
-                Or a list of subset names to filter for.
-            new_result_prefix (str | None): If not None, add a prefix to the new
-                results to distinguish new results from the original TabArena results.
-                Use this, for example, if you re-run a model from TabArena.
-        """
-        results = self.get_results(
-            new_result_prefix=new_result_prefix,
-            use_model_results=use_model_results,
-            fillna=not filter_dataset_fold,
-        )
-
-        return compare_on_tabarena(
-            new_results=results,
-            output_dir=output_dir,
-            filter_dataset_fold=filter_dataset_fold,
-            subset=subset,
-        )
-
-    def get_results(
-        self,
-        new_result_prefix: str | None = None,
-        use_model_results: bool = False,
-        fillna: bool = False,
-    ) -> pd.DataFrame:
-        """
-        Get data to compare results on TabArena leaderboard.
-            Args:
-                new_result_prefix (str | None): If not None, add a prefix to the new
-                    results to distinguish new results from the original TabArena results.
-                    Use this, for example, if you re-run a model from TabArena.
-        """
-        use_model_results = self.method_metadata.method_type != "config" or use_model_results
-
-        if use_model_results:
-            df_results = self.model_results
-        else:
-            df_results = self.hpo_results
-
-        if new_result_prefix is not None:
-            for col in ["method", "config_type", "ta_name", "ta_suite"]:
-                df_results[col] = new_result_prefix + df_results[col]
-
-        if fillna:
-            df_results = self.fillna_results_on_tabarena(df_results=df_results)
-
-        return df_results
-
-    @classmethod
-    def fillna_results_on_tabarena(cls, df_results: pd.DataFrame) -> pd.DataFrame:
-        tabarena_context = TabArenaContext()
-        fillna_method = "RF (default)"
-        fillna_method_name = "RandomForest"
-
-        df_fillna = tabarena_context.load_results_paper(methods=[fillna_method_name])
-        df_fillna = df_fillna[df_fillna["method"] == fillna_method]
-        assert not df_fillna.empty
-
-        # FIXME: Nick: After imputing: ta_name, ta_suite, config_type, etc. are incorrect,
-        #  need to use original, not filled values
-        #  This doesn't impact the evaluation, but could introduce bugs in future if we use these columns
-        #  Fixing this is do-able, but requires some complex pandas tricks, so I haven't had time to implement it yet
-        return TabArenaContext.fillna_metrics(
-            df_to_fill=df_results,
-            df_fillna=df_fillna,
-        )
